requirements.txt
Package                    Version
-------------------------- ----------
aniso8601                  8.0.0
attrs                      19.3.0
bcrypt                     3.1.7
beautifulsoup4             4.9.1
bert-extractive-summarizer 0.4.2
bert-tensorflow            1.0.1
blis                       0.4.1
boto3                      1.13.16
botocore                   1.16.16
bs4                        0.0.1
catalogue                  1.0.0
certifi                    2020.4.5.1
cffi                       1.14.0
chardet                    3.0.4
click                      7.1.2
cymem                      2.0.3
docutils                   0.15.2
en-core-web-sm@ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
filelock                   3.0.12
Flask                      1.1.2
Flask-Bcrypt               0.7.1
flask-restplus             0.13.0
Flask-Script               2.0.6
Flask-Testing              0.8.0
ftfy                       5.7
future                     0.18.2
idna                       2.9
importlib-metadata         1.6.0
itsdangerous               1.1.0
Jinja2                     2.11.2
jmespath                   0.10.0
joblib                     0.15.1
jsonschema                 3.2.0
MarkupSafe                 1.1.1
more-itertools             8.3.0
murmurhash                 1.0.2
nltk                       3.5
numpy                      1.18.4
ordered-set                4.0.1
packaging                  20.4
Pillow                     7.1.2
pip                        20.1.1
plac                       1.1.3
pluggy                     0.13.1
preshed                    3.0.2
probableparsing            0.0.1
protobuf                   3.12.1
py                         1.8.1
pycparser                  2.20
pyparsing                  2.4.7
pyrsistent                 0.16.0
pytest                     5.4.2
python-crfsuite            0.9.7
python-dateutil            2.8.1
pytokenizations            0.4.10
pytz                       2020.1
regex                      2020.5.14
requests                   2.23.0
s3transfer                 0.3.3
sacremoses                 0.0.43
scikit-learn               0.23.1
scipy                      1.4.1
sentencepiece              0.1.91
setuptools                 46.4.0
six                        1.15.0
soupsieve                  2.0.1
spacy                      2.2.4
spacy-stanza               0.2.1
spacy-transformers         0.6.0
srsly                      1.0.2
stanza                     1.0.1
textblob                   0.15.3
thinc                      7.4.0
threadpoolctl              2.0.0
tokenizers                 0.5.2
torch                      1.5.0
torchcontrib               0.0.2
torchvision                0.6.0
tqdm                       4.46.0
transformers               2.5.1
urllib3                    1.25.9
usaddress                  0.5.10
waitress                   1.4.3
wasabi                     0.6.0
wcwidth                    0.1.9
Werkzeug                   0.16.1
zipp                       3.1.0
